browser-service/pyproject.toml
[project]
name = "browser-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "yenslife", email = "77geo5rge6@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "browser-use==0.12.6",
    "fastapi>=0.135.3",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.42.0",
]

[project.scripts]
browser-service = "browser_service:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/browser_service"]
